+++ host/class/uvc/usb_host_uvc/include/esp_private/uvc_esp_video.h
@@ -14,7 +14,7 @@ extern "C" {
 
 /**
  * This API provides access to advanced UVC features.
- * It is intended for use with esp_vide component.
+ * It is intended for use with esp_video component.
  */
 
 typedef struct {